requirements.txt
aiohttp==3.7.4.post0
git+https://github.com/Infiniticity/akinator.py
async_timeout==3.0.1
asyncpraw==7.5.0
certifi==2021.10.8
git+https://github.com/Rapptz/discord.py
googletrans==3.1.0a0
humanfriendly==10.0
motor==2.5.1
numerize==0.12
Pillow==9.0.1
pymongo==3.12.3
dnspython
lyricsgenius
requests==2.26.0
utils==1.0.1
wavelink
python-dotenv